requirements.txt
fastapi==0.88.0
pydantic==1.10.3
pymongo==4.0.2
python-decouple==3.6
uvicorn==0.14.0
dnspython==2.1.0
passlib==1.7.4
bcrypt==3.2.0
jwt==1.3.1
python-jose==3.3.0
python-multipart==0.0.5
email-validator==1.3.0